+++ tests/internal_fn.rs
@@ -1,11 +1,11 @@
-use rhai::{Engine, EvalAltResult};
+use rhai::{Engine, EvalAltResult, INT};
 
 #[test]
 fn test_internal_fn() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
-    assert_eq!(engine.eval::<i64>("fn addme(a, b) { a+b } addme(3, 4)")?, 7);
-    assert_eq!(engine.eval::<i64>("fn bob() { return 4; 5 } bob()")?, 4);
+    assert_eq!(engine.eval::<INT>("fn addme(a, b) { a+b } addme(3, 4)")?, 7);
+    assert_eq!(engine.eval::<INT>("fn bob() { return 4; 5 } bob()")?, 4);
 
     Ok(())
 }
@@ -15,7 +15,7 @@ fn test_big_internal_fn() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
     assert_eq!(
-        engine.eval::<i64>(
+        engine.eval::<INT>(
             r"
                 fn mathme(a, b, c, d, e, f) {
                     a - b * c + d * e - f
