+++ tests/number_literals.rs
@@ -1,10 +1,10 @@
-use rhai::{Engine, EvalAltResult};
+use rhai::{Engine, EvalAltResult, INT};
 
 #[test]
 fn test_number_literal() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
-    assert_eq!(engine.eval::<i64>("65")?, 65);
+    assert_eq!(engine.eval::<INT>("65")?, 65);
 
     Ok(())
 }
@@ -13,8 +13,8 @@ fn test_number_literal() -> Result<(), EvalAltResult> {
 fn test_hex_literal() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
-    assert_eq!(engine.eval::<i64>("let x = 0xf; x")?, 15);
-    assert_eq!(engine.eval::<i64>("let x = 0xff; x")?, 255);
+    assert_eq!(engine.eval::<INT>("let x = 0xf; x")?, 15);
+    assert_eq!(engine.eval::<INT>("let x = 0xff; x")?, 255);
 
     Ok(())
 }
@@ -23,8 +23,8 @@ fn test_hex_literal() -> Result<(), EvalAltResult> {
 fn test_octal_literal() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
-    assert_eq!(engine.eval::<i64>("let x = 0o77; x")?, 63);
-    assert_eq!(engine.eval::<i64>("let x = 0o1234; x")?, 668);
+    assert_eq!(engine.eval::<INT>("let x = 0o77; x")?, 63);
+    assert_eq!(engine.eval::<INT>("let x = 0o1234; x")?, 668);
 
     Ok(())
 }
@@ -33,9 +33,9 @@ fn test_octal_literal() -> Result<(), EvalAltResult> {
 fn test_binary_literal() -> Result<(), EvalAltResult> {
     let mut engine = Engine::new();
 
-    assert_eq!(engine.eval::<i64>("let x = 0b1111; x")?, 15);
+    assert_eq!(engine.eval::<INT>("let x = 0b1111; x")?, 15);
     assert_eq!(
-        engine.eval::<i64>("let x = 0b0011_1100_1010_0101; x")?,
+        engine.eval::<INT>("let x = 0b0011_1100_1010_0101; x")?,
         15525
     );
